Minor changes to avoid warning for signed/unsigned comparisons

+++ base/Genotype.h
@@ -46,7 +46,7 @@ namespace ea {
 
         std::string toString() {
             std::string str = "";
-            for (int i = 0; i < size; ++i)
+            for (size_t i = 0; i < size; ++i)
                 std::visit([&str](const auto& x) { str += " "; str += std::to_string(x); }, genes[i]);
             return str;
         }
